Fixed a bug in the Ant script that were causing tests to refuse to run sometimes because of an (erroneous) Hypersonic error.
As of this build, we have eliminated the need to sign JSPWiki JARs for 95% of most users' deployments. The only case where you need to sign them, now, is when you wish to create a JVM-wide, consolidated security policy. We have retained an Ant target "signjar" for that purpose. If you are upgrading from an earlier version of JSPWiki and are using a customized policy, however, you MUST MUST MUST remove any "signedBy jspwiki" and "keystore" references, otherwise JSPWiki might act odd.


git-svn-id: https://svn.apache.org/repos/asf/incubator/jspwiki/trunk@626456 13f79535-47bb-0310-9956-ffa450edef68